split out rakefile_csvs
Separate Rakefile for the task of making the Term CSVs
