+++ rakefile_common.rb
@@ -5,6 +5,8 @@ require 'pry'
 require 'rake/clean'
 require 'yajl/json_gem'
 
+require_relative 'rakefile_csvs.rb'
+
 Numeric.class_eval { def empty?; false; end }
 
 
@@ -206,111 +208,6 @@ namespace :transform do
   task :check_no_embedded_areas => :ensure_legislature do
     raise "Memberships should not have embedded areas" if @json[:memberships].any? { |m| m.key? :area }
     raise "Memberships must all have legislative_periods" if @json[:memberships].any? { |m| m[:legislative_period_id].to_s.empty? }
-  end
-
-end
-
-
-desc "Build the term-table CSVs"
-task :csvs => ['term_csvs:term_tables']
-
-CLEAN.include('term-*.csv')
-
-namespace :term_csvs do
-
-  require 'csv'
-
-  def persons_twitter(p)
-    if p.key? :contact_details
-      if cd_twitter = p[:contact_details].find { |d| d[:type] == 'twitter' }
-        return cd_twitter[:value].strip
-      end
-    end
-
-    if p.key? 'links'
-      if l_twitter = p[:links].find { |d| d[:note][/twitter/i] }
-        return l_twitter[:url].strip
-      end
-    end
-  end
-
-  # https://twitter.com/tom_watson?lang=en
-  # https://twitter.com/search?q=%23EvelynMEP&src=typd
-
-  def standardised_twitter(t)
-    return if t.to_s.empty?
-    return $1 if t.match /^\@(\w+)$/
-    return $1 if t.match /^(\w+)$/
-    return $1 if t.match %r{(?:www.)?twitter.com/@?(\w+)$}
-
-    # Odd special cases
-    return $1 if t.match %r{twitter.com/search\?q=%23(\w+)}
-    return $1 if t.match %r{twitter.com/#!/https://twitter.com/(\w+)}
-    return $1 if t.match %r{(?:www.)?twitter.com/#!/(\w+)[/\?]?}
-    return $1 if t.match %r{(?:www.)?twitter.com/@?(\w+)[/\/]?}
-    warn "Unknown twitter handle: #{t.to_s.magenta}"
-    return 
-  end
-
-  def persons_facebook(p)
-    (p[:links] || {}).find(->{{url: nil}}) { |d| d[:note] == 'facebook' }[:url]
-  end
-
-  def name_at(p, date)
-    return p[:name] unless date && p.key?(:other_names)
-    historic = p[:other_names].find_all { |n| n.key?(:end_date) } 
-    return p[:name] unless historic.any?
-    at_date = historic.find_all { |n|
-      n[:end_date] >= date && (n[:start_date] || '0000-00-00') <= date
-    }
-    return p[:name] if at_date.empty?
-    raise "Too many names at #{date}: #{at_date}" if at_date.count > 1
-    
-    return at_date.first[:name]
-  end
-
-  task :term_tables => 'ep-popolo-v1.0.json' do
-    @json = JSON.parse(File.read('ep-popolo-v1.0.json'), symbolize_names: true )
-    terms = {}
-
-    data = @json[:memberships].find_all { |m| m.key? :legislative_period_id }.map do |m|
-      person = @json[:persons].find       { |r| (r[:id] == m[:person_id])       || (r[:id].end_with? "/#{m[:person_id]}") }
-      group  = @json[:organizations].find { |o| (o[:id] == m[:on_behalf_of_id]) || (o[:id].end_with? "/#{m[:on_behalf_of_id]}") }
-      house  = @json[:organizations].find { |o| (o[:id] == m[:organization_id]) || (o[:id].end_with? "/#{m[:organization_id]}") }
-      terms[m[:legislative_period_id]] ||= @json[:events].find { |e| e[:id].split('/').last == m[:legislative_period_id].split('/').last }
-
-      if group.nil?
-        puts "No group for #{m}"
-        binding.pry
-        next
-      end
-
-      {
-        id: person[:id].split('/').last,
-        name: name_at(person, m[:end_date] || terms[m[:legislative_period_id]][:end_date]),
-        email: person[:email],
-        twitter: standardised_twitter(persons_twitter(person)),
-        facebook: persons_facebook(person),
-        group: group[:name],
-        group_id: group[:id].split('/').last,
-        area_id: m[:area_id],
-        area: m[:area_id] && @json[:areas].find { |a| a[:id] == m[:area_id] }[:name],
-        chamber: house[:name],
-        term: m[:legislative_period_id].split('/').last,
-        start_date: m[:start_date],
-        end_date: m[:end_date],
-        image: person[:image],
-        gender: person[:gender],
-      }
-    end
-    data.group_by { |r| r[:term] }.each do |t, rs|
-      filename = "term-#{t}.csv"
-      header = rs.first.keys.to_csv
-      rows   = rs.sort_by { |r| [r[:name], r[:start_date].to_s] }.map { |r| r.values.to_csv }
-      csv    = [header, rows].compact.join
-      warn "Creating #{filename}"
-      File.write(filename, csv)
-    end
   end
 
 end
